chore: rename sub packages
