+++ src/loader/pattern.rs
