+++ src/rasterizer/bitmap.rs
